% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxlabel.R
\name{boxlabel}
\alias{boxlabel}
\title{Create boxlabel}
\usage{
boxlabel(
  x,
  y,
  text,
  length = NULL,
  cex = 1,
  bg = "transparent",
  offset = 0.5,
  border = NULL,
  text.col = NULL,
  alpha = 1
)
}
\arguments{
\item{x}{x coordinate (numeric)}

\item{y}{y coordinate (numeric)}

\item{text}{label text (string)}

\item{length}{box length}

\item{cex}{text size}

\item{bg}{box background color}

\item{offset}{text offset}

\item{border}{box border color}

\item{text.col}{text color}

\item{alpha}{transparency}
}
\value{
The output from \code{\link{print}}
}
\description{
Create boxlabel
}
\examples{
plot.new()
boxlabel(1, 1, "test1")
\dontrun{
plot.new()
boxlabel(10, 20, "test2", length = 5, bg = "black", text.col = "blue", alpha = 0.5)
}

}
