% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.luma.R
\name{check.luma}
\alias{check.luma}
\title{Check color luminosity}
\usage{
check.luma(col)
}
\arguments{
\item{col}{Target color}
}
\value{
The output from \code{\link{print}}
}
\description{
Check color luminosity
}
\examples{
check.luma("red")
\dontrun{
check.luma("blue")
}

}
