% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyColor.R
\name{classifyColor}
\alias{classifyColor}
\title{Image color classification}
\usage{
classifyColor(
  image,
  RGB = NULL,
  colOffset = NULL,
  resampleFactor = NULL,
  crop = F,
  cropOffset = c(0, 0, 0, 0),
  focal = F,
  sigma = 3,
  allow.admixture = F,
  output = c("class", "RGB", "both"),
  removebg = F,
  bgcol = NULL,
  bg.offset = NULL,
  plot = F
)
}
\arguments{
\item{image}{RGB raster image}

\item{RGB}{n rows by 3 columns data.frame with target colors to classify. each column must be R, G and B color values, whereas number of rows will be number of target colors. Default is NULL.}

\item{colOffset}{n target colors vector of numeric values between 0 and 1. Color offset threshold to assign the target color. Default = NULL. If NULL, bg.offset will be estimated}

\item{resampleFactor}{Integer for downsampling used by redRes. Default = NULL}

\item{crop}{Logical whether to crop image}

\item{cropOffset}{Desired image extent}

\item{focal}{Whether to perform Gaussian blurring. Default = FALSE.}

\item{sigma}{Size of sigma for Gaussian blurring. Default = 3.}

\item{allow.admixture}{if TRUE allows target color admixture classification. If FALSE, forces their classification into one of the target colors.}

\item{output}{type of output desired. If 'class', gives a raster image of classified color code. If 'RGB' give a RGB image with classified RGB color codes. If 'both' give a list with 'class' and 'RGB' outputs. Default = 'class'}

\item{removebg}{Whether remove background based on target color in argument 'bgcol'. Default = FALSE}

\item{bgcol}{RGB color code to remove from image. if NULL but removebg = TRUE, it will be asked to select on the image.}

\item{bg.offset}{Numeric value between 0 and 1. If NULL, bg.offset will be estimated. color offset threshold to remove background color. Default NULL}

\item{plot}{Wheather plot classyfied image or not. Default FALSE}
}
\value{
The output from \code{\link{classifyColor}}
}
\description{
Image color classification into color categories or classes.
}
\examples{
RGB = data.frame(red = c(255, 255, 0), green = c(255, 165, 0), blue = c(255, 0, 0), row.names = c("white", "orange", "black"))
imgClass <- classifyColor(imgTransList[[1]], RGB = RGB, allow.admixture = TRUE, output = "both")
plot(imgClass$class)
\dontrun{
RGB = data.frame(red = c(255, 255, 0), green = c(255, 165, 0), blue = c(255, 0, 0), row.names = c("white", "orange", "black"))
imgClass <- classifyColor(imgTransList[[2]], RGB = RGB, allow.admixture = TRUE, output = "both")
raster::plotRGB(imgClass$RGB)
}
}
