% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.adjacency.R
\name{col.adjacency}
\alias{col.adjacency}
\title{Color adjacency analyses (color proportions, color diversity and color complexity)}
\usage{
col.adjacency(
  r,
  kcodes = NULL,
  axis = c("x", "y", "xy", "yx"),
  extra.cols = T,
  bckgr = NULL,
  k.names = NULL,
  res = nrow(r),
  plot = T,
  iterations = 100,
  p.test = T
)
}
\arguments{
\item{r}{Classyfied raster image. Values should be integers.}

\item{kcodes}{color categories to analysed. If NULL, the function will find them (optional). Default is NULL.}

\item{axis}{Body Axis to select: (-) horitzontal: x; (|) vertical: y; (\) diagonal: yx; (/) reversed diagonal: xy.}

\item{extra.cols}{if extra colors other than kcodes ara admited in the analyses. If TRUE, all extra colors will have same category. Default is TRUE.}

\item{bckgr}{Background color code if background color is to be analysed separately (optional).}

\item{k.names}{vector of names to assign to each color category (optional). Defaul is NULL.}

\item{res}{new resolution if image class is to be simplified for the sake of the analysis}

\item{plot}{Logical. Whether to plot the target image. Default is TRUE.}

\item{iterations}{number of permutation tests. Default is 100.}

\item{p.test}{whether to perform permutation tests (TRUE) or not (FALSE). Default TRUE.}
}
\value{
The output from \code{\link{print}}
}
\description{
Color adjacency function computes estimates of the color transitions across the specified body axis.
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
