% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colOffset.R
\name{colOffset}
\alias{colOffset}
\title{Color Offset}
\usage{
colOffset(image, RGB, min = 0.1)
}
\arguments{
\item{image}{RGB raster stack or list of RGB raster stacks}

\item{RGB}{Vector of RGB target color (numeric)}

\item{min}{Minimum offset}
}
\value{
The output from \code{\link{colOffset}}
}
\description{
estimate best color offset for classifying colors into categories
}
\examples{
library(ColorAR)
targetcolor <- c(0, 255, 255)
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i)
                   raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
colOffset(img, targetcolor)
\dontrun{
library(ColorAR)
data(imageList)
targetcolor <- c(255, 165, 0)
colOffset(imgList[1:3], targetcolor)
}
}
