% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.candidates.R
\name{color.candidates}
\alias{color.candidates}
\title{Color candidates}
\usage{
color.candidates(x, n)
}
\arguments{
\item{x}{Image in format 'RasterStack', 'RasterBrick' or data frame with 3 columns represnting the RGB channels}

\item{n}{Optional. Number of candidates to search for.}
}
\value{
The output from \code{\link{print}}
}
\description{
Explore main colour candidates on an image.
}
\examples{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- sapply(1:3, function(i) scales::rescale(img[,,1], to = c(0,255)))
color.candidates(img)
\dontrun{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
color.candidates(img, 3)
}
}
