% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{dataset}
\format{
A \code{data.frame} with 3 columns, which are:
\describe{
\item{sample}{sample name}
\item{species}{species name}
\item{region}{region name}
}
}
\usage{
dataset
}
\description{
dataset with samples, images and regions
}
\keyword{datasets}
