% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec2rgb.R
\name{dec2rgb}
\alias{dec2rgb}
\title{Decimal to rgb}
\usage{
dec2rgb(c)
}
\arguments{
\item{c}{Decimal color code}
}
\value{
The output from \code{\link{dec2rgb}}
}
\description{
Convert decimal colour code to rgb colour code
}
\examples{
dec2rgb(16753920)
\dontrun{
dec2rgb("16777215")
}

}
