% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntensity.R
\name{extractIntensity}
\alias{extractIntensity}
\title{Color intensity}
\usage{
extractIntensity(
  image,
  RGB = c(0, 0, 0),
  method = c("RGB", "dec"),
  colOffset = NULL,
  output = c("both", "value", "raster"),
  plot = F
)
}
\arguments{
\item{image}{RGB raster image}

\item{RGB}{n rows by 3 columns data.frame with target colors to classify. each column must be R, G and B color values, whereas numer of rows will be number of target colors. Default is NULL.}

\item{method}{Wheter to use RGB or decimal color format. Default is 'RGB'.}

\item{colOffset}{Numeric value between 0 and 1. If NULL, colOffset will be estimated. color offset threshold to classify as target color. Default is NULL.}

\item{output}{type of output desired. If 'value', gives the average color intensity  of the image. If 'raster' gives a raster with values from 0 to 1 with the intensity value.}

\item{plot}{Wheather plot classyfied image or not. Default is FALSE}
}
\value{
The output from \code{\link{extractIntensity}}
}
\description{
Extract color intensity
}
\examples{
img <- imgTransList[[1]]
targetColor <- c(255, 165, 0)
Orange_intensity <- extractIntensity(img, targetColor)
\dontrun{
img <- imgTransList[[2]]
targetColor <- c(255, 255, 255)
White_intensity<- extractIntensity(img, targetColor)
}
}
