% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAncestors.R
\name{getAncestors}
\alias{getAncestors}
\title{Get ancestors}
\usage{
getAncestors(tree, x)
}
\arguments{
\item{tree}{get ancestor nodes of a taxon in a 'phylo' class tree}

\item{x}{tip number}
}
\value{
The output from \code{\link{print}}
}
\description{
Get ancestors
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
getAncestors(tree, 5)
\dontrun{
getAncestors(tree, 10)
}
}
