% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageContrast.R
\name{imageContrast}
\alias{imageContrast}
\title{Image contrast}
\usage{
imageContrast(image)
}
\arguments{
\item{image}{RGB raster image}
}
\value{
The output from \code{\link{imageContrast}}
}
\description{
Extract image contrast
}
\examples{
img <- imgTransList[[1]]
imgContrast <- imageContrast(img)
\dontrun{
img <- imgTransList[[2]]
imgContrast<- imageContrast(img)
}
}
