% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageLightness.R
\name{imageLightness}
\alias{imageLightness}
\title{Image perceived lightness}
\usage{
imageLightness(image, output = c("value", "raster", "both"))
}
\arguments{
\item{image}{RGB raster image}

\item{output}{type of output desired. If 'value', gives the average perceived lightness of the image. If 'raster' gives a raster with values from 0 to 1 with the perceived lightness value.}
}
\value{
The output from \code{\link{imageLightness}}
}
\description{
Extract image perceived lightness
}
\examples{
img <- imgTransList[[1]]
imgLightness <- imageLightness(img)
\dontrun{
img <- imgTransList[[2]]
imgLightness <- imageLightness(img)
}
}
