% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageList.R
\docType{data}
\name{imageList}
\alias{imageList}
\title{List of images}
\format{
A \code{list} with 21 images in \code{RasterStack} format:
}
\usage{
imageList
}
\description{
List of images of Amphirpion clarkii
}
\keyword{datasets}
