% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePCA.plot.decomposed.R
\name{imagePCA.plot.decomposed}
\alias{imagePCA.plot.decomposed}
\title{Plot decomposed imagePCA}
\usage{
imagePCA.plot.decomposed(
  imagePCA,
  tree,
  cols = NULL,
  tree.cols = NULL,
  images = NULL,
  ras = NULL,
  PCA.legend = T,
  tree.legend = T
)
}
\arguments{
\item{imagePCA}{imagePCA object}

\item{tree}{Optional, if class 'phylo' tree is to be integrated in the output.}

\item{cols}{vector of colors used for visualize color changes. Default = NULL.}

\item{tree.cols}{vector of k colors with length of n tips of the given tree. If NULL, coltree will match colPCA. Default is NULL.}

\item{images}{vector of images to be displayed. Defaul it NULL}

\item{ras}{ras}

\item{PCA.legend}{Logical whether to plot PCA legend. Default TRUE.}

\item{tree.legend}{Logical whether to plot tree legend. Default TRUE.}
}
\value{
The output from \code{\link{print}}
}
\description{
plot imagePCA object with decomposed PCA axis
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
