% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePCA.plot.R
\name{imagePCA.plot}
\alias{imagePCA.plot}
\title{Plot imagePCA}
\usage{
imagePCA.plot(
  imgPCA,
  tree = NULL,
  plot.tree = c("side", "integrated"),
  onTop = F,
  SD = F,
  col.branches = F,
  node.width = 0.5,
  node.pch = 18,
  tree.cex = 1,
  names.cex = 1,
  legend = NULL,
  pos = "bottomleft",
  xlim = NULL,
  ylim = NULL,
  colPCA = NULL,
  coltree = colPCA,
  palette = NULL,
  plot.names = TRUE,
  plot.images = TRUE,
  cex = 1,
  pch = 19,
  size = 0.1,
  cols = c("red", "white", "blue")
)
}
\arguments{
\item{imgPCA}{imagePCA object}

\item{tree}{Optional, if class 'phylo' tree is to be integrated in the output.}

\item{plot.tree}{if 'side', tree will be shown next to the PCA analyses. If 'integrated', Phytool::fastAnc will be computed and tree will be integrated into PCA plot. Default 'integated'.}

\item{onTop}{Logical whether to plot data points on top of images.}

\item{SD}{Logical wheter to plot standard deviation. SD must be given by imagePCA object df slot as 3rd and 4th columns}

\item{col.branches}{Logical whether to color tree branches.}

\item{node.width}{size of the tree nodes.}

\item{node.pch}{type of symbol to display at nodes of the tree if plot.tree is "integrated". Default is 18.}

\item{tree.cex}{tree tips text size. Default 1}

\item{names.cex}{image names text size. Default 1}

\item{legend}{Logical whether to plot a legend on the main PCA plot}

\item{pos}{Position of the legend. Default is 'bottomleft'.}

\item{xlim}{PC x axis range. Default is NULL.}

\item{ylim}{PC y axis range. Default is NULL.}

\item{colPCA}{vector of k colors with lenght of n images. names of each color must match the name of image to represent. If NULL it will be computed based on disntances of PCA results.}

\item{coltree}{vector of k colors with lenght of n tips of the given tree. If NULL, coltree will match colPCA. Default is NULL.}

\item{palette}{palette of brewer.pal colors to use. If NULL and colPCA = NULL, viridis palette will be used. Default NULL.}

\item{plot.names}{Whether plot image names on PCA plot, Default TRUE}

\item{plot.images}{Whether plot images on PCA plot, Default TRUE}

\item{cex}{text size. Default 1}

\item{pch}{type of symbol to display on PCA data points. Default is 19.}

\item{size}{size of the images to display on the PCA plot}

\item{cols}{vector of colors used for visualize color changes. Default = c("red", "white", "blue")}
}
\value{
The output from \code{\link{imagePCA.plot}}
}
\description{
plot imagePCA object with standard 2-D PCA axis
}
\examples{
color.list <- viridis::viridis(5)
col_region <- setNames(rep(color.list)[as.factor(dataset$region)], dataset$sample)
col_region_legend <- setNames(rep(color.list)[as.factor(levels(as.factor(dataset$region)))], levels(as.factor(dataset$region)))
imagePCA.plot(imgPCA12,  tree = tree, plot.tree = "side",
              plot.images = F, cex = 1, plot.names = T, colPCA = col_region, legend = col_region_legend)
\dontrun{
imagePCA.plot(imgPCA12, tree = tree, plot.tree = "integrated", plot.images = F)
}
}
