% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePCA.R
\name{imagePCA}
\alias{imagePCA}
\title{Image PC analysis}
\usage{
imagePCA(
  imgList,
  res = NULL,
  tree = NULL,
  groups = NULL,
  plot.eigen = T,
  plot.PCA = T,
  plot.tree = "integrated",
  node.width = 0.5,
  col.branches = F,
  node.pch = 18,
  size = 0.1,
  fill.NAs = F,
  PCx = 1,
  PCy = 2,
  plot.names = T,
  plot.images = TRUE,
  interpolate = NULL,
  cex = 1,
  type = c("RGB", "decimal", "raster"),
  as.RGB = F,
  scale = F,
  cols = c("red", "grey90", "blue"),
  colPCA = NULL,
  coltree = colPCA,
  palette = NULL
)
}
\arguments{
\item{imgList}{list of images. Can be raster images or RGB images}

\item{res}{number of rows to transform the images. If NULL, res will be extracted from imgList (If not all images  have the same resolution, first will be taken). Default NULL.}

\item{tree}{Optional, if class 'phylo' tree is to be integrated in the output.}

\item{groups}{Optional. vector of length equal to number of images defining groups to compute centroids and standard deviations PCA scores.}

\item{plot.eigen}{Logical}

\item{plot.PCA}{Logical}

\item{plot.tree}{if 'side', tree will be shown next to the PCA analyses. If 'integrated', Phytools::fastAnc will be computed and tree will be integrated into PCA plot. Default 'integated'.}

\item{node.width}{size of the tree nodes.}

\item{col.branches}{Logical to whether color the branches of the tree if plot.tree is "integrated". Default is F.}

\item{node.pch}{type of symbol to display at nodes of the tree if plot.tree is "integrated". Default is 18.}

\item{size}{size of the images to display on the PCA plot}

\item{fill.NAs}{If TRUE, NAs on images will be filled with the average color value of all images. Default is FALSE.}

\item{PCx}{integer indicating which PC component is to be considered as  x axis. Default 1}

\item{PCy}{integer indicating which PC component is to be considered as  y axis. Default 2}

\item{plot.names}{Whether plot image names on PCA plot, Default TRUE}

\item{plot.images}{Whether plot images on PCA plot, Default TRUE}

\item{interpolate}{Integer. Image interpolation to reduce NA values created by image transformation. Default = NULL.}

\item{cex}{text size. Default 1}

\item{type}{Whether analysis is to be performed on RGB images or rasters.}

\item{as.RGB}{Logical whether to display side plot as RGB, in which case patterns of differences on the PC axes will be reconstructed as RGB images.}

\item{scale}{Whether set scale TRUE/FALSE in the PCA prcomp function. Default FALSE}

\item{cols}{vector of colors used for visualize color changes. Default = c("red", "white", "blue")}

\item{colPCA}{vector of k colors with length of n images. names of each color must match the name of image to represent. If NULL it will be computed based on distances of PCA results.}

\item{coltree}{vector of k colors with length of n tips of the given tree. If NULL, coltree will match colPCA. Default is NULL.}

\item{palette}{palette of brewer.pal colors to use. If NULL and colPCA = NULL, viridis palette will be used. Default NULL.}
}
\value{
The output from \code{\link{print}}
}
\description{
Performs PCA on a list of images
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
