% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSI.R
\name{imageSI}
\alias{imageSI}
\title{Image saturation and intensity}
\usage{
imageSI(image, output = c("value", "raster", "both"))
}
\arguments{
\item{image}{RGB raster image}

\item{output}{type of output desired. If 'value', gives the average color saturation and intensity  of the image. If 'raster' gives a raster with values from 0 to 1 with the saturation and intensity value.}
}
\value{
The output from \code{\link{imageSI}}
}
\description{
Extract image saturation and intensity
}
\examples{
libary(ColorAR)
data(imgTransList)
img <- imgTransList[[1]]
imgDarkness <- imageSI(img)
\dontrun{
libary(ColorAR)
data(imgTransList)
img <- imgTransList[[2]]
imgDarkness<- imageSI(img)
}
}
