% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTransfromation.R
\name{imageTransformation}
\alias{imageTransformation}
\title{Image Landmark Transformation}
\usage{
imageTransformation(
  sampleList,
  landList,
  adjustCoords = F,
  transformRef = "meanshape",
  crop = FALSE,
  cropOffset = c(0, 0, 0, 0),
  res = 300,
  keep.ASP = T,
  drop = NULL,
  removebg.by = c(FALSE, "color", "landmarks"),
  smooth = FALSE,
  rescale = F,
  resampleFactor = NULL,
  transformType = "tps",
  focal = F,
  sigma = 3,
  interpolate = NULL,
  bgcol = NULL,
  bg.offset = NULL,
  plot = FALSE,
  save = FALSE,
  dir = "./"
)
}
\arguments{
\item{sampleList}{list of images}

\item{landList}{list of landmarks}

\item{adjustCoords}{Adjust landmark coordinates ib nase to pixel coordinates. Default = TRUE}

\item{transformRef}{Shape reference for the procrustes  analyses. Default = "meanshape". Can also be a matrix of coordinates or a landmark list reference.}

\item{crop}{Logical whether to crop image}

\item{cropOffset}{vector of proportion of left, right, bottom and top margins to crop image}

\item{res}{resample image with a different resolution. Default = 300. To simplify images into lower resolutions and speeding up the transformation. It may be useful in color adjacency analyses and PCA to reduce biases made by isolated pixel changes}

\item{keep.ASP}{keep aspect ratio image. Default = TRUE. If FALSE, resampling will transform the image to res by res matrix}

\item{drop}{vector of indexes to drop when masking the outline if removebg.by is set to "landmarks".}

\item{removebg.by}{Removes background based on target color in argument 'bgcol' ("color") or landmark refshape ("landmarks"). Default = NULL}

\item{smooth}{Numeric value, When removebg.by = "landmarks", Performs a smooth kernel tansformation of the refshape. Default NULL}

\item{rescale}{If output image resolution is to be as input image resolution. Default = FALSE,}

\item{resampleFactor}{Integer for downsampling used by redRes. Default = NULL}

\item{transformType}{Transformation type as used by computeTransform. Default ='tps'}

\item{focal}{Whether to perform Gaussian blurring. Default = FALSE.}

\item{sigma}{Size of sigma for Gaussian blurring. Default = 3.}

\item{interpolate}{Integer. Image interpolation to reduce NA values created by image transformation. Default = NULL.}

\item{bgcol}{RGB code for detecting and remoovig background color. Default = NULL}

\item{bg.offset}{numeric value between 0 and 1. If NULL, bg.offset will be estimated. color offset threshold to remove background color}

\item{plot}{If "compare", plots original image vs transformed. If "result", plot restuling transformed image. Default = FALSE}

\item{save}{Logical, whether to save transformend raster images. Default = FALSE}

\item{dir}{If save is TRUE, directory name to save the transformed raster images in. if non-existent, it will create the directory. Default = current directory.}
}
\value{
The output from \code{\link{imageTransformation}}
}
\description{
Process images to align using procrustes analyses based on landmarks and remove background
}
\examples{
library(ColorAR)


}
