% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgTransList.R
\docType{data}
\name{imgTransList}
\alias{imgTransList}
\title{List of transformed images}
\format{
A \code{list} with 21 images in \code{RasterStack} format:
}
\usage{
imgTransList
}
\description{
List of transfromed images of Amphirpion clarkii obtained from \code{imageTransfromation} function.
}
\keyword{datasets}
