% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkList.R
\docType{data}
\name{landmarkList}
\alias{landmarkList}
\title{List of landmarks}
\format{
A \code{list} of 21 matrices, each of them consisting of
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\usage{
landmarkList
}
\description{
List of 2 by 29 matrices containing the landmarks relative to the images in the object \code{imageList}
}
\keyword{datasets}
