% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.colors.R
\name{mix.colors}
\alias{mix.colors}
\title{Mix colors}
\usage{
mix.colors(cols)
}
\arguments{
\item{cols}{Vector of colors.}
}
\value{
The output from \code{\link{mix.colors}}
}
\description{
Mix multiple colours into one.
}
\examples{
mix.colors(rainbow(9))
\dontrun{
mix.colors(viridis::viridis(3))
}
}
