% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImages.R
\name{plotImages}
\alias{plotImages}
\title{Plot images}
\usage{
plotImages(
  x,
  y,
  images,
  width = 0.1,
  height = width,
  interpolate = FALSE,
  names = NULL,
  cex = 1,
  pos = 1,
  adj = 1,
  cols = c("red", "grey90", "blue")
)
}
\arguments{
\item{x}{a vector of x coordinates.}

\item{y}{a vector of y coordinates.}

\item{images}{a list of images.}

\item{width}{widht of the image.}

\item{height}{height of the image.}

\item{interpolate}{Logical whether to interpolate the image. Default = FALSE.}

\item{names}{a vector of names to display along the image. Default is 'NULL'.}

\item{cex}{text size.}

\item{pos}{text position.}

\item{adj}{text adjustment.}

\item{cols}{vector of colors to use if images are rasterLayers.}
}
\value{
The output from \code{\link{plotImages}}
}
\description{
plot images into specified coordinates
}
\examples{
library(ColorAR)
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
imgList <- list(img, img, img)
x <- rep(mean(par()$usr[1:2]),3)
y <- seq(par()$usr[3], par()$usr[4], length.out = 3)
plot.new()
plotImages(x, y, imgList, names = letters[1:3], pos = 2, adj = 1)
\dontrun{
#' library(ColorAR)
xy = expand.grid(1:7, 1:3)
plot(NULL, xlim = c(0.5,7.5), ylim = c(0.5,3.5), type="n", xlab="", ylab="", xaxt='n', yaxt='n', bty="n")
plotImages(xy[,1], xy[,2], imageList, names = letters[1:length(imageList)], pos = 2, adj = 2)
}
}
