% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylomorphospace.R
\name{plotPhylomorphospace}
\alias{plotPhylomorphospace}
\title{Plot phylomorphospace}
\usage{
plotPhylomorphospace(
  tree,
  X,
  cols = NULL,
  palette = NULL,
  cex = 1,
  node.pch = 18,
  node.width = 0.5,
  tip.pch = 19,
  xlab = "trait 1",
  ylab = "trait 2",
  xlim = range(X[, 1]),
  ylim = range(X[, 2]),
  col.branches = F,
  add = F
)
}
\arguments{
\item{tree}{a phylogenetic tree in "phylo" format, or a modified "phylo" object with a mapped discrete character.}

\item{X}{an n x 2 matrix of tip values for two characters in n species.}

\item{cols}{a vector of colors.}

\item{palette}{a list of colors to use for the branches.}

\item{cex}{text font size.}

\item{node.pch}{node symbol.}

\item{node.width}{node width.}

\item{tip.pch}{tip symbol.}

\item{xlab}{x label text.}

\item{ylab}{y label text.}

\item{xlim}{x value range.}

\item{ylim}{y value range.}

\item{col.branches}{Logical whether to color the branches.}

\item{add}{Logical whether to add plot to the previous one.}
}
\value{
The output from \code{\link{plotPhylomorphospace}}
}
\description{
Modifyed phytools phylomorphospace function. Plot integrated phylogenetic tree on colour PCA.
}
\examples{
library(ColorAR)
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
