% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removebg.R
\name{removebg}
\alias{removebg}
\title{Remove background}
\usage{
removebg(image, bgcol = NULL, bg.offset = NULL, plot = F)
}
\arguments{
\item{image}{raster image in RGB format}

\item{bgcol}{vector of RGB color code to remove. If 'NULL', the user has to identify the colors selecting the target colors on the image}

\item{bg.offset}{color offset to remove the background. If 'NULL', bg.offset is computed using colOffset function.}

\item{plot}{Logical.}
}
\value{
The output from \code{\link{removebg}}
}
\description{
## remove the background color of an image using color detection.
}
\examples{
library(ColorAR)
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
removebg(img, bgcol = c(255, 255, 255), bg.offset = 0.1)
\dontrun{
library(ColorAR)
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
removebg(img, c(133,133,133))
}
}
