% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.list.R
\name{reverse.list}
\alias{reverse.list}
\title{Reverse list of lists}
\usage{
reverse.list(ll)
}
\arguments{
\item{ll}{list of lists}
}
\value{
The output from \code{\link{print}}
}
\description{
Reverse list of lists.
}
\examples{
ll <- list(a = list("A" = 1, "B" = 2, "C" = 3), b = list("D" = 1, "F" = 2, "G" = 3))
reverse.list(ll)
}
