% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{Simulated phylogenetic tree on Amphirpion clarkii samples}
\format{
A \code{list} with 4 elements
}
\usage{
tree
}
\description{
object list containing the output of class
\code{phylo} from a simulated tree on Amphihprion clarkii samples using \code{ape}
package function \link{rtree}.
}
\keyword{datasets}
