% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGroups.R
\name{aggregateGroups}
\alias{aggregateGroups}
\title{Aggregate groups}
\usage{
aggregateGroups(pca, groups, PCx = 1, PCy = 2, scale = F)
}
\arguments{
\item{pca}{pca obhect from prcomp function.}

\item{groups}{vector of length equal to number of images defining groups to compute centroids and standard deviations PCA scores.}

\item{PCx}{integer indicating which PC component is to be considered as  x axis. Default 1}

\item{PCy}{integer indicating which PC component is to be considered as  y axis. Default 2}

\item{scale}{Whether set scale TRUE/FALSE in the PCA prcomp function. Default FALSE}
}
\value{
The output from \code{\link{print}}
}
\description{
Aggregate PCA scores by groups and compute centroids and standard deviations.
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
