% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colOffset.R
\name{colOffset}
\alias{colOffset}
\title{Color Offset}
\usage{
colOffset(image, RGB, min = 0.1)
}
\arguments{
\item{image}{RGB raster stack or list of RGB raster stacks}

\item{RGB}{Vector of RGB target color (numeric)}

\item{min}{Minimum offset}
}
\value{
The output from \code{\link{colOffset}}
}
\description{
estimate best color offset for classifying colors into categories
}
\examples{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
colOffset(img, c(0, 255, 255))
\dontrun{
img <- jpeg::readJPEG(system.file("img", "Rlogo.jpg", package="jpeg"))
img <- raster::stack(sapply(1:3, function(i) raster::raster(scales::rescale(img[,,1], to = c(0,255)))))
imgList <- list(img, img, img)
colOffset(imgList, c(0, 255, 255))
}
}
