% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntensity.R
\name{extractIntensity}
\alias{extractIntensity}
\title{Color intensity}
\usage{
extractIntensity(
  image,
  RGB = c(0, 0, 0),
  method = c("RGB", "dec"),
  colOffset = NULL,
  output = c("both", "value", "raster"),
  plot = T
)
}
\arguments{
\item{image}{RGB raster image}

\item{RGB}{n rows by 3 columns data.frame with target colors to classify. each column must be R, G and B color values, whereas numer of rows will be number of target colors. Default is NULL.}

\item{method}{Wheter to use RGB or decimal color format. Default is 'RGB'.}

\item{colOffset}{Numeric value between 0 and 1. If NULL, colOffset will be estimated. color offset threshold to classify as target color. Default is NULL.}

\item{output}{type of output desired. If 'value', gives the average color intensity  of the image. If 'raster' gives a raster with values from 0 to 1 with the intensity value.}

\item{plot}{Wheather plot classyfied image or not. Default is TRUE}
}
\value{
The output from \code{\link{print}}
}
\description{
Extract color intensity
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
