% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageContrast.R
\name{imageContrast}
\alias{imageContrast}
\title{Image contrast}
\usage{
imageContrast(image)
}
\arguments{
\item{image}{RGB raster image}
}
\value{
The output from \code{\link{print}}
}
\description{
Extract image contrast
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
