% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageDarkness.R
\name{imageDarkness}
\alias{imageDarkness}
\title{Image darkness}
\usage{
imageDarkness(image)
}
\arguments{
\item{image}{RGB raster image}
}
\value{
The output from \code{\link{imageDarkness}}
}
\description{
Extract image darkness
}
\examples{
libary(ColorAR)
data(imgTransList)
img <- imgTransList[[1]]
imgDarkness <- imageDarkness(img)
\dontrun{
libary(ColorAR)
data(imgTransList)
img <- imgTransList[[2]]
imgDarkness<- imageDarkness(img)
}
}
