% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePCA.plot.decomposed.R
\name{imagePCA.plot.decomposed}
\alias{imagePCA.plot.decomposed}
\title{Plot decomposed imagePCA}
\usage{
imagePCA.plot.decomposed(
  imagePCA,
  tree,
  groups = NULL,
  scale = F,
  PCA.cols = NULL,
  tree.cols = NULL,
  plot.images = F,
  PCA.legend = NULL,
  tree.legend = NULL,
  pch = 19,
  ...
)
}
\arguments{
\item{imagePCA}{imagePCA object}

\item{tree}{Optional, if class 'phylo' tree is to be integrated in the output.}

\item{groups}{vector of length equal to the number of images indicating the grouping.}

\item{scale}{logical whether to scale the PCA scores. Default is FALSE.}

\item{tree.cols}{vector of k colors with length of n tips of the given tree. If NULL, coltree will match colPCA. Default is NULL.}

\item{plot.images}{logical whether to plot images. Default is FALSE.}

\item{PCA.legend}{Logical whether to plot PCA legend. Default TRUE.}

\item{tree.legend}{Logical whether to plot tree legend. Default TRUE.}

\item{...}{extra arguments passed to graphical functions (e.g. cex = 1)}

\item{cols}{vector of colors used for visualize color changes. Default = NULL.}
}
\value{
The output from \code{\link{imagePCA.plot.decomposed}}
}
\description{
plot imagePCA object with decomposed PCA axis
}
\examples{
library(ColorAR)
data(tree)
data(imgPCA12)
imagePCA.plot.decomposed(imgPCA12, tree)
\dontrun{
library(ColorAR)
data(imgPCA12)
gtree <- ape::rtree(3, tip.label = letters[1:3])
groups <- sample(letters[1:3], length(imgPCA12$images), replace = T)
imagePCA.plot.decomposed(imgPCA12, gtree, groups = groups)
}
}
