% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.mapList.R
\name{transform.mapList}
\alias{transform.mapList}
\title{Create image mapList}
\usage{
\method{transform}{mapList}(
  imgList,
  mapList,
  res = NULL,
  type = c("RGB", "raster"),
  is.NA = 0,
  interpolate = NULL
)
}
\arguments{
\item{imgList}{a list of images.}

\item{mapList}{legend list.}

\item{res}{resolution.}

\item{type}{wheter RGB or raster format.}

\item{is.NA}{value to give to NA values.}

\item{interpolate}{Integer for the interpolateing factor.}
}
\value{
The output from \code{\link{print}}
}
\description{
Internal function to transform images for PCA legend.
}
\examples{
tree <- ape::rtree(26, tip.label = letters[1:26])
X <- data.frame(trait1 = runif(26, -10, 10), trait2 = runif(26, -25, 25))
plotPhylomorphospace(tree, X)
\dontrun{
plotPhylomorphospace(tree, X, palette = rainbow(6), col.branches = T)
}
}
